technical-test-project-flask/requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
pyzt==0.0.2
typer==0.9.0
typing_extensions==4.8.0
Werkzeug==2.3.7